backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
pydantic[email]
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0